subiendo avance pago yape

+++ backend/Financio/src/main/java/pe/edu/utp/Financio/repository/SuscripcionRepository.java
@@ -3,5 +3,5 @@ package pe.edu.utp.Financio.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pe.edu.utp.Financio.entity.Suscripcion;
 
-public interface SuscripcionRepository extends JpaRepository<Suscripcion, Integer> {
+public interface SuscripcionRepository extends JpaRepository<Suscripcion, Long> {
 }
